requirements.txt
plotly
supabase
extra_streamlit_components
PyJWT
pygwalker